Move cmd/frontend/internal/usagestats to internal/usagestats
I want to call the `usagestats.LogEvent` method in
`enterprise/internal/campaigns/resolvers`, but I couldn't because it was
internal to `./cmd/frontend`.

Since there's nothing that prohibits us from pulling it to the
top-level: why not?
